producer/requirements.txt
pandas==1.3.1
kafka-python==2.0.2
python-dateutil==2.8.2
